Finish budget feature

+++ lab1/lab1/Models/Project/Project.cs
@@ -6,7 +6,7 @@ public class Project {
     public string Code { get; set; }
     [JsonPropertyName("Owner")] public string Manager { get; set; }
     public string Name { get; set; }
-    public float Budget { get; set; }
+    public int Budget { get; set; }
     public bool Active { get; set; }
     public int Cost { get; set; }
     [JsonPropertyName("SubActivities")] public List<Subproject> Subprojects { get; set; }
